+++ composeApp/src/androidMain/kotlin/com/nuvio/app/features/player/PlayerEngine.android.kt
@@ -78,6 +78,7 @@ actual fun PlatformPlayerSurface(
     sourceResponseHeaders: Map<String, String>,
     streamType: String?,
     useYoutubeChunkedPlayback: Boolean,
+    @Suppress("UNUSED_PARAMETER") isAnimeContent: Boolean,
     modifier: Modifier,
     playWhenReady: Boolean,
     resizeMode: PlayerResizeMode,